more frontend for questions sidebar
