+++ src/app/(dashboard)/layout.tsx
@@ -1,6 +1,7 @@
 import { AppSidebar } from "@/components/app-sidebar";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
-import { ChatPanel, ChatProvider } from "@/features/chat/components/chat";
+import { ChatPanel } from "@/features/chat/components/chat-panel";
+import { ChatProvider } from "@/features/chat/components/context";
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
@@ -8,8 +9,9 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
       <SidebarProvider>
         <AppSidebar />
         <SidebarInset className="bg-accent/20">{children}</SidebarInset>
+        {/*TODO: make chat panel resizable*/}
+        <ChatPanel />
       </SidebarProvider>
-      <ChatPanel />
     </ChatProvider>
   );
 };
